# swagger_server/models/phrase.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Phrase(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, id: int=None, original: str=None, translation: str=None, example_recording: str=None):  # noqa: E501
        """Phrase - a model defined in Swagger

        :param id: The id of this Phrase.  # noqa: E501
        :type id: int
        :param original: The original of this Phrase.  # noqa: E501
        :type original: str
        :param translation: The translation of this Phrase.  # noqa: E501
        :type translation: str
        :param example_recording: The example_recording of this Phrase.  # noqa: E501
        :type example_recording: str
        """
        self.swagger_types = {
            'id': int,
            'original': str,
            'translation': str,
            'example_recording': str
        }

        self.attribute_map = {
            'id': 'id',
            'original': 'original',
            'translation': 'translation',
            'example_recording': 'exampleRecording'
        }
        self._id = id
        self._original = original
        self._translation = translation
        self._example_recording = example_recording

    @classmethod
    def from_dict(cls, dikt) -> 'Phrase':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Phrase of this Phrase.  # noqa: E501
        :rtype: Phrase
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> int:
        """Gets the id of this Phrase.


        :return: The id of this Phrase.
        :rtype: int
        """
        return self._id

    @id.setter
    def id(self, id: int):
        """Sets the id of this Phrase.


        :param id: The id of this Phrase.
        :type id: int
        """

        self._id = id

    @property
    def original(self) -> str:
        """Gets the original of this Phrase.


        :return: The original of this Phrase.
        :rtype: str
        """
        return self._original

    @original.setter
    def original(self, original: str):
        """Sets the original of this Phrase.


        :param original: The original of this Phrase.
        :type original: str
        """

        self._original = original

    @property
    def translation(self) -> str:
        """Gets the translation of this Phrase.


        :return: The translation of this Phrase.
        :rtype: str
        """
        return self._translation

    @translation.setter
    def translation(self, translation: str):
        """Sets the translation of this Phrase.


        :param translation: The translation of this Phrase.
        :type translation: str
        """

        self._translation = translation

    @property
    def example_recording(self) -> str:
        """Gets the example_recording of this Phrase.


        :return: The example_recording of this Phrase.
        :rtype: str
        """
        return self._example_recording

    @example_recording.setter
    def example_recording(self, example_recording: str):
        """Sets the example_recording of this Phrase.


        :param example_recording: The example_recording of this Phrase.
        :type example_recording: str
        """

        self._example_recording = example_recording